+++ forensicfit/machine_learning/data/data_generation.py
@@ -12,14 +12,13 @@ from functools import partial
 import pandas as pd
 from sklearn.model_selection import train_test_split
 from PIL import Image, ImageOps
-
 from sklearn.model_selection import KFold
 
+from forensicfit.utils import ROOT
 
 Image.MAX_IMAGE_PIXELS = 933120000
-SCRATCH_DIR = f"{os.sep}users{os.sep}lllang{os.sep}SCRATCH"
-PROJECT_DIR = os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
-DATA_DIR = f"{PROJECT_DIR}{os.sep}data"
+
+DATA_DIR = f"{ROOT}{os.sep}data"
 
 @dataclass
 class DataGenerationConfig:
