+++ forensicfit/machine_learning/download_models.py
@@ -2,14 +2,12 @@ import os
 import shutil
 import gdown
 
-PROJECT_DIR = os.path.dirname(os.path.dirname(__file__))
+from forensicfit.utils import ROOT
 
 def download_models():
 
-    machine_learning_dir = os.path.join(PROJECT_DIR,'machine_learning')
-
-    output = f"{machine_learning_dir}{os.sep}models.zip"
-    to = f"{machine_learning_dir}{os.sep}"
+    output = f"{ROOT}{os.sep}models.zip"
+    to = f"{ROOT}{os.sep}"
     url  = f'https://drive.google.com/drive/folders/1viez7q5TmwkdbcwceS4dWxeBsM8bDg0X'
