+++ forensicfit/utils/__init__.py
@@ -7,3 +7,15 @@ if HAS_OPENCV:
     from . import image_tools
 from . import plotter
 from .general import copy_doc
+
+
+import os
+import json
+from glob import glob
+import logging.config
+from pathlib import Path
+
+
+# Other Constants
+FILE = Path(__file__).resolve()
+ROOT = str(FILE.parents[1])  # forensicfit
